Active Defense/requirements.txt
flask
flask-socketio
requests
watchdog
dnslib
scapy
python-dotenv
